+++ src/domain/collaboration/callout/post/CalloutPostCreatedSubscription.graphql
