+++ src/domain/collaboration/callout/post/CalloutPostsSubscription.graphql
@@ -1,0 +1,10 @@
+query CalloutPostsSubscription($calloutId: UUID!) {
+  lookup {
+    callout(ID: $calloutId) {
+      id
+      posts {
+        ...ContributeTabPost
+      }
+    }
+  }
+}
