+++ src/domain/collaboration/callout/post/PostCallout.tsx
@@ -1,15 +1,9 @@
-import React, { forwardRef, useMemo, useState } from 'react';
+import React, { forwardRef, Ref, useMemo, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useInView } from 'react-intersection-observer';
 import CalloutLayout, { CalloutLayoutProps } from '../../CalloutBlock/CalloutLayout';
 import ScrollableCardsLayout from '../../../../core/ui/card/cardsLayout/ScrollableCardsLayout';
 import PostCreationDialog from '../../post/PostCreationDialog/PostCreationDialog';
-import {
-  PostCardFragmentDoc,
-  useCreatePostFromContributeTabMutation,
-} from '../../../../core/apollo/generated/apollo-hooks';
-import { usePostCreatedOnCalloutSubscription } from '../usePostCreatedOnCalloutSubscription';
-import { CalloutState, CreatePostOnCalloutInput, TagsetType } from '../../../../core/apollo/generated/graphql-schema';
+import { CalloutState, CreatePostOnCalloutInput } from '../../../../core/apollo/generated/graphql-schema';
 import CreateCalloutItemButton from '../CreateCalloutItemButton';
 import { buildPostUrl } from '../../../../main/routing/urlBuilders';
 import PostCard, { PostCardPost } from './PostCard';
@@ -18,18 +12,25 @@ import { gutters } from '../../../../core/ui/grid/utils';
 import CalloutBlockFooter from '../../CalloutBlock/CalloutBlockFooter';
 import useCurrentBreakpoint from '../../../../core/ui/utils/useCurrentBreakpoint';
 import PageContentBlock from '../../../../core/ui/content/PageContentBlock';
-import { useCombinedRefs } from '../../../shared/utils/useCombinedRefs';
 
 export type OnCreateInput = Omit<CreatePostOnCalloutInput, 'calloutID'>;
 
 interface PostCalloutProps extends BaseCalloutViewProps {
   callout: CalloutLayoutProps['callout'];
+  posts: PostCardPost[] | undefined;
+  loading: boolean;
+  creatingPost: boolean;
+  onCreatePost: (post: OnCreateInput) => Promise<{ nameID: string } | undefined>;
 }
 
-const PostCallout = forwardRef<HTMLDivElement, PostCalloutProps>(
+const PostCallout = forwardRef<Element, PostCalloutProps>(
   (
     {
       callout,
+      posts,
+      loading,
+      creatingPost,
+      onCreatePost,
       canCreate = false,
       spaceNameId,
       challengeNameId,
@@ -46,98 +47,98 @@ const PostCallout = forwardRef<HTMLDivElement, PostCalloutProps>(
     const closeCreateDialog = () => setPostDialogOpen(false);
     const navigate = useNavigate();
 
-    const { ref: intersectionObserverRef, inView } = useInView({
-      delay: 500,
-      trackVisibility: true,
-      triggerOnce: true,
-    });
-
-    const { subscriptionEnabled, posts, loading } = usePostCreatedOnCalloutSubscription({
-      spaceNameId,
-      calloutId: callout.id,
-      challengeNameId,
-      opportunityNameId,
-      skip: !inView,
-    });
-
-    const [createPost, { loading: isCreatingPost }] = useCreatePostFromContributeTabMutation({
-      update: (cache, { data }) => {
-        if (subscriptionEnabled || !data) {
-          return;
-        }
-
-        const { createPostOnCallout } = data;
-
-        const calloutRefId = cache.identify({
-          __typename: 'Callout',
-          id: callout.id,
-        });
-
-        if (!calloutRefId) {
-          return;
-        }
-
-        cache.modify({
-          id: calloutRefId,
-          fields: {
-            posts(existingPosts = []) {
-              const newPostRef = cache.writeFragment({
-                data: createPostOnCallout,
-                fragment: PostCardFragmentDoc,
-                fragmentName: 'PostCard',
-              });
-              return [...existingPosts, newPostRef];
-            },
-          },
-        });
-      },
-    });
-
-    const onCreate = async (post: OnCreateInput) => {
-      const { data } = await createPost({
-        variables: {
-          postData: {
-            calloutID: callout.id,
-            profileData: {
-              displayName: post.profileData.displayName,
-              description: post.profileData.description,
-            },
-            tags: post.tags,
-            type: post.type,
-            visualUri: post.visualUri,
-          },
-        },
-        optimisticResponse: {
-          createPostOnCallout: {
-            __typename: 'Post',
-            id: '',
-            nameID: '',
-            profile: {
-              id: '',
-              displayName: post.profileData.displayName,
-              description: post.profileData?.description,
-              visual: {
-                id: '-1',
-                name: '',
-                uri: post.visualUri ?? '',
-              },
-              tagset: {
-                id: '-1',
-                name: 'default',
-                tags: [],
-                allowedValues: [],
-                type: TagsetType.Freeform,
-              },
-            },
-            type: post.type,
-          },
-        },
-      });
-
-      const nameID = data?.createPostOnCallout.nameID;
-
-      return nameID ? { nameID } : undefined;
-    };
+    // const { ref: intersectionObserverRef, inView } = useInView({
+    //   delay: 500,
+    //   trackVisibility: true,
+    //   triggerOnce: true,
+    // });
+    //
+    // const { subscriptionEnabled, posts, loading } = usePostCreatedOnCalloutSubscription({
+    //   spaceNameId,
+    //   calloutId: callout.id,
+    //   challengeNameId,
+    //   opportunityNameId,
+    //   skip: !inView,
+    // });
+    //
+    // const [createPost, { loading: isCreatingPost }] = useCreatePostFromContributeTabMutation({
+    //   update: (cache, { data }) => {
+    //     if (subscriptionEnabled || !data) {
+    //       return;
+    //     }
+    //
+    //     const { createPostOnCallout } = data;
+    //
+    //     const calloutRefId = cache.identify({
+    //       __typename: 'Callout',
+    //       id: callout.id,
+    //     });
+    //
+    //     if (!calloutRefId) {
+    //       return;
+    //     }
+    //
+    //     cache.modify({
+    //       id: calloutRefId,
+    //       fields: {
+    //         posts(existingPosts = []) {
+    //           const newPostRef = cache.writeFragment({
+    //             data: createPostOnCallout,
+    //             fragment: PostCardFragmentDoc,
+    //             fragmentName: 'PostCard',
+    //           });
+    //           return [...existingPosts, newPostRef];
+    //         },
+    //       },
+    //     });
+    //   },
+    // });
+    //
+    // const onCreate = async (post: OnCreateInput) => {
+    //   const { data } = await createPost({
+    //     variables: {
+    //       postData: {
+    //         calloutID: callout.id,
+    //         profileData: {
+    //           displayName: post.profileData.displayName,
+    //           description: post.profileData.description,
+    //         },
+    //         tags: post.tags,
+    //         type: post.type,
+    //         visualUri: post.visualUri,
+    //       },
+    //     },
+    //     optimisticResponse: {
+    //       createPostOnCallout: {
+    //         __typename: 'Post',
+    //         id: '',
+    //         nameID: '',
+    //         profile: {
+    //           id: '',
+    //           displayName: post.profileData.displayName,
+    //           description: post.profileData?.description,
+    //           visual: {
+    //             id: '-1',
+    //             name: '',
+    //             uri: post.visualUri ?? '',
+    //           },
+    //           tagset: {
+    //             id: '-1',
+    //             name: 'default',
+    //             tags: [],
+    //             allowedValues: [],
+    //             type: TagsetType.Freeform,
+    //           },
+    //         },
+    //         type: post.type,
+    //       },
+    //     },
+    //   });
+    //
+    //   const nameID = data?.createPostOnCallout.nameID;
+    //
+    //   return nameID ? { nameID } : undefined;
+    // };
 
     const postNames = useMemo(() => posts?.map(x => x.profile.displayName) ?? [], [posts]);
 
@@ -159,14 +160,12 @@ const PostCallout = forwardRef<HTMLDivElement, PostCalloutProps>(
 
     const isMobile = breakpoint === 'xs';
 
-    const containerRef = useCombinedRefs(null, ref, intersectionObserverRef);
-
     return (
       <>
-        <PageContentBlock ref={containerRef} disablePadding disableGap {...blockProps}>
+        <PageContentBlock ref={ref as Ref<HTMLDivElement>} disablePadding disableGap {...blockProps}>
           <CalloutLayout callout={callout} contributionsCount={contributionsCount} {...calloutLayoutProps}>
             <ScrollableCardsLayout
-              items={loading || !inView ? [undefined, undefined] : posts ?? []}
+              items={loading ? [undefined, undefined] : posts ?? []}
               deps={[spaceNameId, challengeNameId, opportunityNameId]}
               createButton={!isMobile && createButton}
               maxHeight={gutters(22)}
@@ -181,7 +180,7 @@ const PostCallout = forwardRef<HTMLDivElement, PostCalloutProps>(
         <PostCreationDialog
           open={postDialogOpen}
           onClose={closeCreateDialog}
-          onCreate={onCreate}
+          onCreate={onCreatePost}
           postNames={postNames}
           calloutDisplayName={callout.profile.displayName}
           spaceNameId={spaceNameId!}
@@ -189,7 +188,7 @@ const PostCallout = forwardRef<HTMLDivElement, PostCalloutProps>(
           opportunityNameId={opportunityNameId}
           calloutId={callout.id}
           postTemplate={callout.postTemplate}
-          isCreating={isCreatingPost}
+          creating={creatingPost}
         />
       </>
     );
