+++ src/domain/collaboration/callout/post/PostsOnCalloutFragment.graphql
