+++ src/domain/collaboration/callout/post/useCalloutPostCreatedSubscription.ts
@@ -1,10 +1,10 @@
-import { CalloutPostCreatedDocument } from '../../../core/apollo/generated/apollo-hooks';
+import { CalloutPostCreatedDocument } from '../../../../core/apollo/generated/apollo-hooks';
 import {
   PostsOnCalloutFragment,
   CalloutPostCreatedSubscription,
   CalloutPostCreatedSubscriptionVariables,
-} from '../../../core/apollo/generated/graphql-schema';
-import createUseSubscriptionToSubEntityHook from '../../../core/apollo/subscriptions/useSubscriptionToSubEntity';
+} from '../../../../core/apollo/generated/graphql-schema';
+import createUseSubscriptionToSubEntityHook from '../../../../core/apollo/subscriptions/useSubscriptionToSubEntity';
 
 const useCalloutPostCreatedSubscription = createUseSubscriptionToSubEntityHook<
   PostsOnCalloutFragment,
