+++ src/domain/collaboration/callout/useCallouts/CalloutQueries.graphql
@@ -35,105 +35,6 @@ query Callouts(
   }
 }
 
-query SpaceCalloutPostsSubscription($spaceNameId: UUID_NAMEID!, $calloutId: UUID_NAMEID!) {
-  space(ID: $spaceNameId) {
-    id
-    collaboration {
-      id
-      callouts(IDs: [$calloutId]) {
-        id
-        posts {
-          ...ContributeTabPost
-        }
-      }
-    }
-  }
-}
-
-query ChallengeCalloutPostsSubscription(
-  $spaceNameId: UUID_NAMEID!
-  $challengeNameId: UUID_NAMEID!
-  $calloutId: UUID_NAMEID!
-) {
-  space(ID: $spaceNameId) {
-    id
-    challenge(ID: $challengeNameId) {
-      id
-      collaboration {
-        id
-        callouts(IDs: [$calloutId]) {
-          id
-          posts {
-            ...ContributeTabPost
-          }
-        }
-      }
-    }
-  }
-}
-
-query OpportunityCalloutPostsSubscription(
-  $spaceNameId: UUID_NAMEID!
-  $opportunityNameId: UUID_NAMEID!
-  $calloutId: UUID_NAMEID!
-) {
-  space(ID: $spaceNameId) {
-    id
-    opportunity(ID: $opportunityNameId) {
-      id
-      collaboration {
-        id
-        callouts(IDs: [$calloutId]) {
-          id
-          posts {
-            ...ContributeTabPost
-          }
-        }
-      }
-    }
-  }
-}
-
-fragment PrivilegesOnCollaboration on Collaboration {
-  id
-  authorization {
-    id
-    myPrivileges
-  }
-}
-query PrivilegesOnSpaceCollaboration($spaceNameId: UUID_NAMEID!) {
-  space(ID: $spaceNameId) {
-    id
-    collaboration {
-      ...PrivilegesOnCollaboration
-    }
-  }
-}
-
-query PrivilegesOnChallengeCollaboration($spaceNameId: UUID_NAMEID!, $challengeNameId: UUID_NAMEID!) {
-  space(ID: $spaceNameId) {
-    id
-    challenge(ID: $challengeNameId) {
-      id
-      collaboration {
-        ...PrivilegesOnCollaboration
-      }
-    }
-  }
-}
-
-query PrivilegesOnOpportunityCollaboration($spaceNameId: UUID_NAMEID!, $opportunityNameId: UUID_NAMEID!) {
-  space(ID: $spaceNameId) {
-    id
-    opportunity(ID: $opportunityNameId) {
-      id
-      collaboration {
-        ...PrivilegesOnCollaboration
-      }
-    }
-  }
-}
-
 fragment CollaborationWithCallouts on Collaboration {
   id
   authorization {
